+++ R/sw-dropdown.R
@@ -4,27 +4,21 @@
 #' Create a dropdown menu
 #'
 #' @param ... List of tag to be displayed into the dropdown menu.
-#' @param style Character. if \code{default} use Bootstrap button (like an \code{actionButton}), else use an
-#' \code{\link{actionBttn}}, see argument \code{style} (in \code{\link{actionBttn}} documentation) for possible values.
-#' @param icon An icon to appear on the button.
-#' @param status Add a class to the buttons, you can use Bootstrap status like 'info', 'primary', 'danger', 'warning' or 'success'.
-#'  Or use an arbitrary strings to add a custom class, e.g. : with \code{status = 'myClass'}, buttons will have class \code{btn-myClass}.
-#' @param size Size of the button : default, lg, sm, xs.
-#' @param label Label to appear on the button. If circle = TRUE and tooltip = TRUE, label is used in tooltip.
-#' @param tooltip Put a tooltip on the button, you can customize tooltip with \code{tooltipOptions}.
+#' @inheritParams actionBttn
+#' @param status Color of the button, see [actionBttn()].
+#' @param tooltip Put a tooltip on the button, you can customize tooltip with [tooltipOptions()].
 #' @param right Logical. The dropdown menu starts on the right.
 #' @param up Logical. Display the dropdown menu above.
 #' @param width Width of the dropdown menu content.
-#' @param animate Add animation on the dropdown, can be logical or result of \code{animateOptions}.
+#' @param animate Add animation on the dropdown, can be logical or result of [animateOptions()].
 #' @param inputId Optional, id for the button, the button act like an \code{actionButton},
 #' and you can use the id to toggle the dropdown menu server-side.
 #'
 #' @details
-#' This function is similar to \code{dropdownButton} but don't use Bootstrap, so you can put \code{pickerInput} in it.
+#' This function is similar to [dropdownButton()] but don't use Bootstrap, so you can use [pickerInput()] in it.
 #' Moreover you can add animations on the appearance / disappearance of the dropdown with animate.css.
 #'
-#' @seealso \code{\link{animateOptions}} for animation, \code{\link{tooltipOptions}} for tooltip and
-#' \code{\link{actionBttn}} for the button.
+#' @seealso [dropMenu()] for a more robust alternative.
 #'
 #' @importFrom htmltools validateCssUnit tagList singleton tags tagAppendChild
 #'
@@ -97,10 +91,20 @@
 #' shinyApp(ui = ui, server = server)
 #'
 #' }
-dropdown <- function(..., style = "default", status = "default",
-                     size = "md", icon = NULL, label = NULL, tooltip = FALSE,
-                     right = FALSE, up = FALSE, width = NULL, animate = FALSE,
-                     inputId = NULL) {
+dropdown <- function(...,
+                     style = "default",
+                     status = "default",
+                     size = "md",
+                     icon = NULL,
+                     label = NULL,
+                     tooltip = FALSE,
+                     right = FALSE,
+                     up = FALSE,
+                     width = NULL,
+                     animate = FALSE,
+                     inputId = NULL,
+                     block = FALSE,
+                     no_outline = TRUE) {
 
 
   if (is.null(inputId)) {
@@ -111,10 +115,12 @@ dropdown <- function(..., style = "default", status = "default",
 
   # Dropdown content
   dropcontent <- htmltools::tags$div(
-    id=contentId, class="sw-dropdown-content animated",
-    class=if(up) "sw-dropup-content", class=if(right) "sw-dropright-content",
-    style=if(!is.null(width)) paste("width:", htmltools::validateCssUnit(width)),
-    htmltools::tags$div(class="sw-dropdown-in", ...)
+    id = contentId,
+    class = "sw-dropdown-content animated",
+    class = if(up) "sw-dropup-content",
+    class = if(right) "sw-dropright-content",
+    style = if(!is.null(width)) paste("width:", htmltools::validateCssUnit(width)),
+    htmltools::tags$div(class = "sw-dropdown-in", ...)
   )
   # Button
   if (style == "default") {
@@ -130,9 +136,14 @@ dropdown <- function(..., style = "default", status = "default",
     )
   } else {
     btn <- actionBttn(
-      inputId = inputId, label = label,
-      icon = icon, style = style,
-      color = status, size = size
+      inputId = inputId,
+      label = label,
+      icon = icon,
+      style = style,
+      color = status,
+      size = size,
+      block = block,
+      no_outline = no_outline
     )
   }